requirements.txt
flask
python-dotenv
pymongo
dnspython
numpy
pandas
scikit-learn
tldextract
